# Reduce JsFunction::call overhead (via smallvec)