--- a/Cargo.toml
+++ b/Cargo.toml
@@ -22,6 +22,7 @@ semver = "0.9"
 cslice = "0.2"
 semver = "0.9.0"
 neon-runtime = { version = "=0.4.2", path = "crates/neon-runtime" }
+smallvec = "1.4.2"
 
 [target.'cfg(target_os = "windows")'.dependencies]
 winapi = { version = "0.3.9", features = ["minwindef", "libloaderapi", "ntdef"] }