@kjvalencik 2020-09-27
This seems like a solid improvement. We are currently driving towards completing the N-API. I would be open to a breaking change for the `napi-runtime` feature.

Since this will always need to be collected into a `Vec` (the n-api function requires an array pointer), I think it would be better to take that directly. I expect most users don't send an iterator. Taking a slice (`&mut [T]`) is probably best because a user could use a `Vec` or in most cases an array.

Given the small performance gain, I would prefer to discuss the API further before bringing in a new dependency. @dherman What do you think?

@sbillig One reason to bring in `smallvec` is that a similar improvement could  be applied to `fn argument`.