@dherman 2020-09-27
I think we should be open to this idea, and certainly N-API is an opportunity for a breaking change. I'm always in favor of being thoughtful about taking on dependencies but `smallvec` is well-tested and well-maintained, so I'm not too worried about that. But it would also be good to figure out just how disruptive the cost of migration would be, what the API and implementation would look like for the N-API runtime, and what the performance difference is there. This should probably have an RFC so we can understand the API change.