@dherman 2020-09-27
But I forgot to say first of all, thank you @sbillig for this idea and contribution! Function calls are so central to Neon that managing their overhead is important.