@kjvalencik 2020-09-28
@dherman I created an issue and gave it the `v1.0` label. https://github.com/neon-bindings/neon/issues/608

In my opinion, it would be rare for a user to construct arguments from an iterator. Most uses will have a fixed number of arguments. Switching to `AsRef<[T]>` will allow us to avoid heap allocation in most places without introducing `smallvec` or copying with the cost of requiring users to manually call `.collect()` if they *are* using an iterator.