@dherman 2020-09-28
@kjvalencik That's awesome, thanks for writing it up! I misunderstood how similar this would be to the current API; I don't think it needs an RFC. And `AsRef[T]` just seems like a better design than `IntoIterator<Item=T>`, don't you think?