@sbillig 2020-09-28
I agree that AsRef<[T]> would make for a better API, and would likely eliminate a vec allocation on the user's side, perhaps doubling the benefit shown here.

Feel free to close this PR if you'd prefer. I'm about to open another one replacing the vec in the napi-runtime CallContext with a smallvec (as suggested by @kjvalencik (that's what you meant, right?)), which similarly improves calls from js to functions defined in rust.